+++ libraries/arrays/weighting.ts
@@ -1,0 +1,5 @@
+export { getCumulativeWeights } from './src/getCumulativeWeights.ts';
+export { getWeightedIndex } from './src/getWeightedIndex.ts';
+export { pickWeightedIndex } from './src/pickWeightedIndex.ts';
+export { toPickWeightedItem } from './src/toPickWeightedItem.ts';
+export { toPickWeightedItemFromDistribution } from './src/toPickWeightedItemFromDistribution.ts';